requirements.txt
streamlit
pandas
plotly
streamlit-calendar
holidays
openpyxl
pyyaml
google-api-python-client
google-auth
google-auth-oauthlib